Fix store-smb split package
Moves the implementation of this plugin from `o.e.i.store` and misc
other places into `o.e.i.store.smb`.
