+++ plugins/store-smb/src/internalClusterTest/java/org/elasticsearch/index/store/smb/SmbSimpleFsTests.java
@@ -6,19 +6,17 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.store;
+package org.elasticsearch.index.store.smb;
 
 import org.elasticsearch.common.settings.Settings;
 
 
-public class SmbMMapFsTests extends AbstractAzureFsTestCase {
-
+public class SmbSimpleFsTests extends AbstractAzureFsTestCase {
     @Override
     public Settings indexSettings() {
         return Settings.builder()
                 .put(super.indexSettings())
-                .put("index.store.type", "smb_mmap_fs")
+                .put("index.store.type", "smb_simple_fs")
                 .build();
     }
-
 }
