+++ plugins/store-smb/src/test/java/org/elasticsearch/index/store/smb/SmbMMapDirectoryTests.java
@@ -6,12 +6,13 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.store;
+package org.elasticsearch.index.store.smb;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.MMapDirectory;
+import org.elasticsearch.index.store.EsBaseDirectoryTestCase;
 
 public class SmbMMapDirectoryTests extends EsBaseDirectoryTestCase {
